Move initial and boundary setup common code to module
