+++ cefi_kit/src/cefi_kit/grids.py
@@ -1,4 +1,5 @@
 import numpy as np
+import xarray
 
 
 def center_to_outer(center, left=None, right=None):
@@ -28,3 +29,20 @@ def corners(lon, lat):
     assert len(lonc) == len(lon) + 1
     assert len(latc) == len(lat) + 1
     return lonc, latc
+
+
+def vgrid_to_interfaces(vgrid, max_depth=6500.0):
+    if isinstance(vgrid, xarray.DataArray):
+        vgrid = vgrid.data
+    zi = np.concatenate([[0], np.cumsum(vgrid)])
+    zi[-1] = max_depth
+    return zi
+
+
+def vgrid_to_layers(vgrid, max_depth=6500.0):
+    if isinstance(vgrid, xarray.DataArray):
+        vgrid = vgrid.data
+    ints = vgrid_to_interfaces(vgrid, max_depth=max_depth)
+    z = (ints + np.roll(ints, shift=1)) / 2
+    layers = z[1:]
+    return layers
